sct_utils: don't crash if git isn't here (#1718)
We need check_exe() so sct_config was collapsed into sct_utils()
